requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0

# Browser Automation
playwright==1.40.0

# Database
sqlalchemy==2.0.25

# Config
pyyaml==6.0.1

# Utils
python-dotenv==1.0.0
requests==2.31.0

# LLM (optional - for comment generation)
openai==1.8.0
anthropic==0.18.0
